fix: Use the up to date scene shape data to calculate mouse event positions.
The previous version would move the actual shape data of the scene (`shape`) on initialisation into the closure that instantiates mouse events. From then on all mouse events were initialised with this data and their positions calculated based on this data. When the scene size changed, this would go out of sync with the actual scene shape and the position of our mouse cursor would no longer align with the actual position of the system cursor.

The fix is to move a handle into the closure that allows access to the up-to-date scene shape data at any time instead of the actual data.