chore: make bug template give better reports
I would really recommend that you modify the bug report template, such that it doesn't have "description" first and "steps to reproduce" below, but just immediately start from "steps to reproduce"; see e.g. https://github.com/golang/go/blob/master/.github/ISSUE_TEMPLATE (from which I blatantly copied some parts; using good sources is a virtue).

The problem is, when a template starts with "description", I (a reporter) start filling it with freeform and somewhat chaotic, unstructured words basically describing the problem and maybe how I came to it; but then I have to repeat most of this work in "steps to reproduce" which is annoying and discouraging.

On the other hand, starting immediately from "steps to reproduce" (in the Go issue template: "What did you do?", which is a bit more open) helps both you and me, because it immediately softly leads me (or another reporter) to start writing it in a structured way that will most useful to you, and doesn't make me feel like I'm repeating myself soon afterwards.