Catch syntax errors in custom visualisations
Syntax errors in custom visualisations caused panics during IDE startup. Now they are caught and handled like errors during execution of the visualisation code, which writes a log message to the console.

This commit duplicates some code from wasm-bindgen to to allow for an easy fix. Hopefully the necessary feature will be included in wasm-bindgen soon.